Замена типов данных В TaroHistoryService добавлен historyList для сохранения выпавших карт в истории одним запросом в БД Создан файл миграции

+++ src/main/java/scari/corp/taro/entity/TaroCards.java
@@ -18,7 +18,7 @@ public class TaroCards {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Short id;
 
     @Column(nullable = false, unique = true)
     private String nameEn;
@@ -31,7 +31,7 @@ public class TaroCards {
 
     private String suit;
 
-    private Integer number;
+    private Short number;
 
     @Embedded
     private Meanings meanings;
